feature:#59/바텀네비게이션 및 기타 UI 추가
